`config nu` now adds frozen jobs to job table
Description:

`config nu` used to ignore the frozen wait job status response and did not
add processes to the job table when they were frozen.

This PR refactors the PostWaitCallback used in run_external and
allows frozen processes spawned by `config_.rs` to be added
to the job table.

Closes #15389
